Implementation and tests added, docstrings and error tests to go
