+++ src/vivarium/framework/results/__init__.py
@@ -1,1 +1,2 @@
-from vivarium.framework.results.manager import ResultsInterface, ResultsManager
+from vivarium.framework.results.interface import ResultsInterface
+
