+++ src/vivarium/framework/results/manager.py
@@ -1,12 +1,14 @@
 from collections import Counter
-from typing import Callable, List, Union
+from typing import TYPE_CHECKING, Callable, List, Union
 
 import pandas as pd
 
-# TODO: importing Builder causes circular import, fix that in implementation
-# from vivarium.framework.engine import Builder
 from vivarium.framework.event import Event
 from vivarium.framework.results.context import ResultsContext
+
+if TYPE_CHECKING:
+    # Cyclic import
+    from vivarium.framework.engine import Builder
 
 
 class ResultsManager:
@@ -38,7 +40,7 @@ class ResultsManager:
         return self._metrics.copy()
 
     # noinspection PyAttributeOutsideInit
-    def setup(self, builder):
+    def setup(self, builder: "Builder"):
         self.population_view = builder.population.get_view([])
         self.clock = builder.time.clock()
         self.step_size = builder.time.step_size()
@@ -65,6 +67,7 @@ class ResultsManager:
         self.gather_results("collect_metrics", event)
 
     def gather_results(self, event_name: str, event: Event):
+        # TODO: Add implementation rather than ...
         ...
         # population = self._prepare_population(event)
         # for results_group in self._results_context.gather_results(population, event_name):
@@ -130,6 +133,8 @@ class ResultsManager:
         )
 
     def _add_resources(self, target: List[str], target_type: str):
+        if not len(target):
+            return  # do nothing on empty lists
         target = set(target) - {"event_time", "current_time", "step_size"}
         if target_type == "column":
             self._required_columns.update(target)
@@ -154,113 +159,3 @@ class ResultsManager:
         # Shim for now to allow incremental transition to new results system.
         metrics.update(self.metrics)
         return metrics
-
-
-class ResultsInterface:
-    """Builder interface for the results management system.
-
-    The results management system allows users to delegate results production
-    to the simulation framework. This process attempts to roughly mimic the
-    groupby-apply logic commonly done when manipulating :mod:`pandas`
-    DataFrames. The representation of state in the simulation is complex,
-    however, as it includes information both in the population state table
-    and dynamically generated information available from the
-    :class:`value pipelines <vivarium.framework.values.Pipeline>`.
-    Additionally, good encapsulation of simulation logic typically has
-    results production separated from the modeling code into specialized
-    `Observer` components. This often highlights the need for transformations
-    of the simulation state into representations that aren't needed for
-    modeling, but are required for the stratification of produced results.
-
-    The purpose of this interface is to provide controlled access to a results
-    backend by means of the builder object. It exposes methods
-    to register stratifications, set default stratifications, and register
-    results producers. There is a special case for stratifications generated
-    by binning continuous data into categories.
-
-    The expected use pattern would be for a single component to register all
-    stratifications required by the model using :func:`register_default_stratifications`,
-    :func:`register_stratification`, and :func:`register_binned_stratification`
-    as necessary. A “binned stratification” is a stratification special case for
-    the very common situation when a continuous value needs to be binned into
-    categorical bins. The `is_vectorized` argument should be True if the mapper
-    function expects a DataFrame, and False if it expects a row of the DataFrame
-    and should be used by calling df.apply.
-    """
-
-    def __init__(self, manager: ResultsManager) -> None:
-        self._manager = manager
-        self._name = "results_interface"
-
-    @property
-    def name(self) -> str:
-        """The name of this ResultsInterface."""
-        return self._name
-
-    def set_default_stratifications(self, default_stratifications: List[str]):
-        self._manager.set_default_stratifications(default_stratifications)
-
-    # TODO: It is not reflected in the sample code here, but the “when” parameter should be added
-    #  to the stratification registration calls, probably as a List.
-    def register_stratification(
-        self,
-        name: str,
-        categories: List[str],
-        mapper: Callable = None,
-        is_vectorized: bool = False,
-        requires_columns: List[str] = (),
-        requires_values: List[str] = (),
-    ) -> None:
-        """Register quantities to observe.
-
-        Parameters
-        ----------
-        name
-            Name of the of the column created by the stratification.
-        categories
-            List of string values that the mapper is allowed to output.
-        mapper
-            A callable that emits values in `categories` given inputs from columns
-            and values in the `requires_columns` and `requires_values`, respectively.
-        is_vectorized
-            `True` if the mapper function expects a `DataFrame`, and `False` if it
-            expects a row of the `DataFrame` and should be used by calling :func:`df.apply`.
-        requires_columns
-            A list of the state table columns that already need to be present
-            and populated in the state table before the pipeline modifier
-            is called.
-        requires_values
-            A list of the value pipelines that need to be properly sourced
-            before the pipeline modifier is called.
-
-        Returns
-        ------
-        None
-        """
-        ...
-        # self._manager.register_stratification(...)
-
-    def register_binned_stratification(
-        self,
-        target: str,
-        binned_column: str,
-        bins: List = (),
-        labels: List[str] = (),
-        target_type: str = "column",
-        **cut_kwargs,
-    ) -> None:
-
-        self._manager.register_binned_stratification(...)
-
-    def register_observation(
-        self,
-        name: str,
-        pop_filter: str = "",
-        aggregator: Callable[[pd.DataFrame], float] = len,
-        requires_columns: List[str] = None,
-        requires_values: List[str] = None,
-        additional_stratifications: List[str] = (),
-        excluded_stratifications: List[str] = (),
-        when: str = "collect_metrics",
-    ) -> None:
-        self._manager.register_observation(...)
